+++ src/apex/buffer.ts
@@ -17,34 +17,80 @@
 
 */
 import { GConstructor } from 'mixin'
-import Apex from './base'
 import { Events } from './events'
 import { changeValue } from '../utils/input'
+import { Util } from './utils'
+
+enum Selector {
+  EmptyBuffer = '#TOUR_TARGET_EMPTY_BUFFER',
+  EmptyBufferNotTaken = '#TOUR_TARGET_EMPTY_BUFFER:not(.prun-palette.prun-taken)',
+  NewBufferButton = '#TOUR_TARGET_BUTTON_BUFFER_NEW',
+  BufferCMDElement = 'div[class^="TileFrame__cmd"]',
+  TakenClass = 'prun-palette.prun-taken',
+}
 
 export type Buffer = GConstructor<{
   createBuffer(command?: string): Promise<Element | null>
+  closeBuffer(buffer: Element): void
+  closeBufferWithCommand(command: string): void
+  closeAllBuffers(): Promise<void>
 }>
 
-export function Buffer<TBase extends GConstructor<Apex> & Events>(Base: TBase) {
+export function Buffer<TBase extends Util & Events>(Base: TBase) {
   return class Buffer extends Base {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    constructor(...args: any[]) {
+      super(...args)
+
+      this.observer.addListener('add', {
+        match: element => {
+          const nodeMatches = element.matches(Selector.EmptyBuffer)
+          const nodeHadMatchingChildren = element.querySelector(
+            Selector.EmptyBuffer
+          )
+          const match = nodeMatches || !!nodeHadMatchingChildren
+          return match
+        },
+        callback: async element => {
+          const buffer = element.matches(Selector.EmptyBuffer)
+            ? element
+            : element.querySelector(Selector.EmptyBuffer)
+
+          if (!buffer) return
+
+          try {
+            const bufferCMDElement = await this.observer.waitFor(
+              Selector.BufferCMDElement,
+              { within: buffer }
+            )
+            const cmd = bufferCMDElement.textContent
+
+            this.events.emit('new-buffer', buffer, cmd ?? undefined)
+          } catch (error) {
+            this.events.emit('new-buffer', buffer)
+          }
+        },
+      })
+    }
+
     private get newBufferButton(): HTMLButtonElement | null {
-      return document.querySelector('#TOUR_TARGET_BUTTON_BUFFER_NEW')
+      return document.querySelector(Selector.NewBufferButton)
     }
 
     public async createBuffer(command?: string): Promise<Element | null> {
       if (this.newBufferButton === null) return null
 
-      const bufferPromise = this.observer.waitFor('#TOUR_TARGET_EMPTY_BUFFER:not(.prun-palette-taken)')
+      const bufferPromise = this.observer.waitFor(
+        Selector.EmptyBufferNotTaken,
+        { onlyNew: true }
+      )
 
       this.newBufferButton.click()
       const buffer = await bufferPromise
 
-      buffer.classList.add('prun-palette-taken')
+      buffer.classList.add(Selector.TakenClass)
 
-      if (!command) {
-        this.events.emit('new-buffer', buffer)
-        return buffer
-      }
+      if (!command) return buffer
 
       const input = buffer.querySelector('input')
 
@@ -52,12 +98,43 @@ export function Buffer<TBase extends GConstructor<Apex> & Events>(Base: TBase) {
         throw new Error('Could not find input element')
       }
 
-      changeValue(input, command)
+      changeValue(input, command.toUpperCase())
       input.form?.requestSubmit()
 
-      this.events.emit('new-buffer', buffer, command)
-
       return buffer
+    }
+
+    public closeBuffer(buffer: Element) {
+      const closeButton = this.findElementWithContent(buffer, 'x')
+      if (!closeButton || !(closeButton instanceof HTMLElement)) return
+
+      closeButton.click()
+    }
+
+    public closeBufferWithCommand(command: string) {
+      const buffers = document.querySelectorAll(Selector.EmptyBuffer)
+
+      Array.from(buffers)
+        .filter(buffer => {
+          const bufferCMDElement = buffer.querySelector(
+            Selector.BufferCMDElement
+          )
+          if (!bufferCMDElement) return false
+
+          const bufferCMD = bufferCMDElement.textContent?.toUpperCase()
+          return bufferCMD === command.toUpperCase()
+        })
+        .forEach(buffer => {
+          this.closeBuffer(buffer)
+        })
+    }
+
+    public async closeAllBuffers(): Promise<void> {
+      const buffers = document.querySelectorAll(Selector.EmptyBuffer)
+
+      buffers.forEach(buffer => {
+        this.closeBuffer(buffer)
+      })
     }
   }
 }