+++ src/components/paletteInput.ts
@@ -22,23 +22,22 @@ import { paletteListener } from './palette'
 
 const cursorEnd = (el: HTMLElement) => {
   const range = document.createRange()
-  range.selectNodeContents(el);
-  range.collapse(false);
-  const selection = window.getSelection();
-  selection?.removeAllRanges();
-  selection?.addRange(range);
+  range.selectNodeContents(el)
+  range.collapse(false)
+  const selection = window.getSelection()
+  selection?.removeAllRanges()
+  selection?.addRange(range)
 }
-
 
 export default function paletteInput(
   path: string[] = [],
-  current: string = '',
-  bestMatch?: string,
+  current = '',
+  bestMatch?: string
 ) {
   const input = div(current)
   input
     .att$('contenteditable', 'true')
-    .onKeyDown$((e) => {
+    .onKeyDown$(e => {
       if (e.key === 'Tab') {
         e.preventDefault()
         return paletteListener.emit('complete', input.innerText)
@@ -64,18 +63,21 @@ export default function paletteInput(
   input.on('mount', focus)
 
   const bestMatchSuffix = bestMatch
-    ? bestMatch.startsWith(current) ? bestMatch.slice(current.length) : ` · ${bestMatch}`
+    ? bestMatch.startsWith(current)
+      ? bestMatch.slice(current.length)
+      : ` · ${bestMatch}`
     : undefined
 
   const container = div(
-    ...path.map((p) => div(p.concat(' ·')).att$('class', 'palette-input-path')),
+    ...path.map(p =>
+      div(p.concat(' ·')).att$('class', 'prun-palette prun-input-path')
+    ),
     input,
-    bestMatchSuffix && div(bestMatchSuffix).att$('class', 'palette-input-best-match')
-  ).att$('class', 'palette-input')
+    bestMatchSuffix &&
+      div(bestMatchSuffix).att$('class', 'prun-alette-input-best-match')
+  ).att$('class', 'prun-palette prun-input')
 
   container.onClick$(() => focus())
 
   return container
 }
-
-